backend/test/test_profile.py
from fastapi.testclient import TestClient
from backend.schemas import ProfileBase, ProfileOnlyRead
from main import app

client = TestClient(app)


def test_get_profile(mocker):
    mock_arg = ProfileBase(
        profile_id="1234",
        profile_name="Test Profile",
        ip_range="192.168.1.0-255",
        last_scan="2024-12-26 20:08:50",
        created_at="2024-12-26 20:08:50",
        updated_at="2024-12-26 20:08:50",
    )

    mocker.patch("backend.service.persistence.db_profile.get_profile", return_value=mock_arg)

    mock_profile = ProfileOnlyRead(
        profile_id="1234",
        profile_name="Test Profile",
        ip_range="192.168.1.0-255",
        last_scan="2024-12-26 20:08:50",
        created_at="2024-12-26 20:08:50",
        updated_at="2024-12-26 20:08:50",
    )
    response = client.get("/profile/1234")
    assert response.status_code == 200
    assert mock_profile.model_dump().get("profile_id") == response.json().get("profile_id")
    assert mock_profile.model_dump().get("profile_name") == response.json().get("profile_name")
    # fails to compare dates properly
    # assert response.json() == mock_profile.model_dump()


def test_get_profiles(mocker):

    mocker.patch(
        "backend.service.persistence.db_profile.get_profiles",
        return_value=[
            ProfileOnlyRead(
                profile_id="1234",
                profile_name="Test Profile",
                profile_description=None,
                ip_range="192.168.1.0-255",
                last_scan="2024-12-26 20:08:50",
                created_at="2024-12-26 20:08:50",
                updated_at="2024-12-26 20:08:50",
            ),
            ProfileOnlyRead(
                profile_id="1234",
                profile_name="Test Profile",
                profile_description=None,
                ip_range="192.168.1.0-255",
                last_scan="2024-12-26 20:08:50",
                created_at="2024-12-26 20:08:50",
                updated_at="2024-12-26 20:08:50",
            ),
        ],
    )

    response = client.get("/profiles/")
    assert response.status_code == 200
    assert len(response.json()) == 2
    assert response.json() == [
        {
            "profile_id": "1234",
            "profile_name": "Test Profile",
            "profile_description": None,
            "ip_range": "192.168.1.0-255",
            "last_scan": "2024-12-26T20:08:50",
            "created_at": "2024-12-26T20:08:50",
            "updated_at": "2024-12-26T20:08:50",
        },
        {
            "profile_id": "1234",
            "profile_name": "Test Profile",
            "profile_description": None,
            "ip_range": "192.168.1.0-255",
            "last_scan": "2024-12-26T20:08:50",
            "created_at": "2024-12-26T20:08:50",
            "updated_at": "2024-12-26T20:08:50",
        },
    ]


def test_get_profiles_latest(mocker):

    mocker.patch(
        "backend.service.persistence.db_profile.get_profiles",
        return_value=[],
    )

    response = client.get("/profiles/latest/1")
    assert response.status_code == 200
    assert len(response.json()) == 0
    assert response.json() == []


def test_scan_profile(mocker):

    mocker.patch(
        "backend.service.profile_scan.ProfileScanService.scan_profile",
        return_value=ProfileBase(
            profile_id="1234",
            profile_name="Test Profile",
            profile_description=None,
            ip_range="192.168.1.0-255",
            last_scan="2024-12-26 20:08:50",
            created_at="2024-12-26 20:08:50",
            updated_at="2024-12-26 20:08:50",
        ),
    )

    response = client.get("/profile/1234/scan")
    assert response.status_code == 200
